Dotted chart

Create a dotted chart to view all events in a glance

dotted_chart(eventlog, x, sort, color, units, ...)

# S3 method for eventlog
dotted_chart(eventlog, x = c("absolute", "relative",
  "relative_week", "relative_day"), sort = c("start", "end", "duration",
  "start_week", "start_day"), color = NULL, units = c("weeks", "days",
  "hours", "mins", "secs"), ...)

# S3 method for grouped_eventlog
dotted_chart(eventlog, x = c("absolute",
  "relative", "relative_week", "relative_day"), sort = c("start", "end",
  "duration", "start_week", "start_day"), color = NULL, units = c("weeks",
  "days", "hours", "mins", "secs"), ...)

idotted_chart(eventlog, plotly = FALSE)

iplotly_dotted_chart(eventlog)

plotly_dotted_chart(eventlog)

Arguments

eventlog
Eventlog object

x
Value for plot on x-axis: absolute time or relative time (since start, since start of week, since start of day)

sort
Ordering of the cases on y-axis: start, end or duration

color
Optional, variable to use for coloring dots. Default is the activity identifier. Use NA for no colors.

units
Time units to use on x-axis in case of relative time.

...
Deprecated arguments

plotly
Return plotly object

Methods (by class)

eventlog: Dotted chart for event log

grouped_eventlog: Dotted chart for grouped event log