Resource Matrix

Construct a resource matrix, showing how work is handed over

resource_matrix(log, type, eventlog = deprecated())

# S3 method for eventlog
resource_matrix(
  log,
  type = c("absolute", "relative", "relative-antecedent", "relative-consequent"),
  eventlog = deprecated()
)

# S3 method for activitylog
resource_matrix(
  log,
  type = c("absolute", "relative", "relative-antecedent", "relative-consequent"),
  eventlog = deprecated()
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

type

The type of resource matrix, which can be absolulte, relative, relative_antecedent or relative_consequent. Absolute will return a matrix with absolute frequencies, relative will return global relative frequencies for all antecedent-consequent pairs. Relative_antecedent will return relative frequencies within each antecendent, i.e. showing the relative proportion of consequents within each antecedent. Relative_consequent will do the reverse.

eventlog

Deprecated. Use log instead.

Methods (by class)

eventlog: Resource matrix of event log

activitylog: Resource matrix of activity log

Examples

if (FALSE) {
library(eventdataR)
data(patients)
precedence_matrix(patients)
}