Trace explorer

Explore traces, ordered by relative trace frequency

trace_explorer(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  .abbreviate = T,
  show_labels = T,
  label_size = 3,
  scale_fill = processmapR:::scale_fill_discrete_bupaR(),
  raw_data = F,
  eventlog = deprecated()
)

# S3 method for eventlog
trace_explorer(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  .abbreviate = T,
  show_labels = T,
  label_size = 3,
  scale_fill = processmapR:::scale_fill_discrete_bupaR(),
  raw_data = F,
  eventlog = deprecated()
)

# S3 method for activitylog
trace_explorer(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  .abbreviate = T,
  show_labels = T,
  label_size = 3,
  scale_fill = processmapR:::scale_fill_discrete_bupaR(),
  raw_data = F,
  eventlog = deprecated()
)

plotly_trace_explorer(
  eventlog,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  .abbreviate = T,
  show_labels = T,
  label_size = 3,
  scale_fill = processmapR:::scale_fill_discrete_bupaR(),
  raw_data = F
)

Arguments

log

log: Object of class log or derivatives (grouped_log, eventlog, activitylog, etc.).

coverage

The percentage coverage of the trace to explore. Default is 20% most (in)frequent

n_traces

Instead of setting coverage, you can set an exact number of traces. Should be an integer larger than 0.

type

Frequent traces first, or infrequent traces first?

coverage_labels

Change the labels to be shown on the right of the process variants. These can be relative frequency (default), absolute, or cumulative.

.abbreviate

If TRUE, abbreviate activity labels

show_labels

If False, activity labels are not shown.

label_size

Font size of labels

scale_fill

Set color scale

raw_data

Return raw data

eventlog

Deprecated. Use log instead.

Methods (by class)

eventlog: Trace explorer eventlog

activitylog: Trace explorer activity log